relocate controller dialogs into its own folder
