+++ CMakeLists.txt
@@ -97,10 +97,10 @@ file (GLOB CXBXR_HEADER_GUIv1
  "${CXBXR_ROOT_DIR}/src/common/input/Button.h"
  "${CXBXR_ROOT_DIR}/src/common/input/EmuDevice.h"
  "${CXBXR_ROOT_DIR}/src/common/input/InputWindow.h"
+ "${CXBXR_ROOT_DIR}/src/gui/controllers/DlgDukeControllerConfig.h"
  "${CXBXR_ROOT_DIR}/src/gui/DlgAbout.h"
  "${CXBXR_ROOT_DIR}/src/gui/DlgAudioConfig.h"
  "${CXBXR_ROOT_DIR}/src/gui/DlgInputConfig.h"
- "${CXBXR_ROOT_DIR}/src/gui/DlgDukeControllerConfig.h"
  "${CXBXR_ROOT_DIR}/src/gui/DlgEepromConfig.h"
  "${CXBXR_ROOT_DIR}/src/gui/DlgLoggingConfig.h"
  "${CXBXR_ROOT_DIR}/src/gui/DlgNetworkConfig.h"
@@ -234,10 +234,10 @@ file (GLOB CXBXR_SOURCE_GUIv1
  "${CXBXR_ROOT_DIR}/src/common/input/Button.cpp"
  "${CXBXR_ROOT_DIR}/src/common/input/EmuDevice.cpp"
  "${CXBXR_ROOT_DIR}/src/common/input/InputWindow.cpp"
+ "${CXBXR_ROOT_DIR}/src/gui/controllers/DlgDukeControllerConfig.cpp"
  "${CXBXR_ROOT_DIR}/src/gui/DlgAbout.cpp"
  "${CXBXR_ROOT_DIR}/src/gui/DlgAudioConfig.cpp"
  "${CXBXR_ROOT_DIR}/src/gui/DlgInputConfig.cpp"
- "${CXBXR_ROOT_DIR}/src/gui/DlgDukeControllerConfig.cpp"
  "${CXBXR_ROOT_DIR}/src/gui/DlgEepromConfig.cpp"
  "${CXBXR_ROOT_DIR}/src/gui/DlgLoggingConfig.cpp"
  "${CXBXR_ROOT_DIR}/src/gui/DlgNetworkConfig.cpp"
