+++ src/gui/controllers/DlgDukeControllerConfig.h
